Resolves 6 conflicts between the Kokkos porting work on this branch and the
incoming small_changes/master development:

- cmake defaults: keep this branch's re-enabled KOKKOS regression tests
  (ablation, surf_collide models, adsorb) and take the incoming exclusions
  for subsonic/mflow emit tests (pending their Kokkos port) and vtk.
- surf_collide.h: keep kokkos_random() hook; take uncopy-member removal,
  bigint counters, and persurf_wrapper().
- fix_emit_surf.h: keep virtual subsonic_* hooks for the /kk subclass; take
  the incoming mflow/window members; make mflow_grid() virtual for the
  upcoming Kokkos port.
- collide_vss_kokkos.{cpp,h}: interleave both sides' additions - the
  incoming transient-subcell (TAS) device implementation, MCF Poisson
  attempt counts, bigint statistics counters, and E-disposal rework
  alongside this branch's multigroup/ambipolar-multigroup collision kernels
  and per-grid gas-tally support. The group-pair attempt_collision_kokkos
  overload gains the MCF branch and the (double) npairs overflow fix to
  match the CPU version. The collisions() dispatch keeps this branch's
  corrected ngas_tally polarity and routes subcell decks through
  collisions_one_subcell (gas-tally instantiations wired next); subcell
  with multiple groups errors out, mirroring the CPU restriction.
- update_kokkos.cpp: keep this branch's typed surf-tally copy lists and
  deferred setup_surf_tally_copies(); take the incoming unconditional
  unused-slot filling for boundary-tally copies and the uncopy removal.

Also removes all remaining uses of the deleted KKCopy uncopy flag from the
Kokkos classes added on this branch (surf_collide adiabatic/cll/td/
impulsive destructors, gas-tally/react-surf/isurf compute copy
constructors), matching the new KKCopy destructor-restore semantics.

+++ src/compute_react_isurf_grid.cpp
@@ -355,8 +355,8 @@ void ComputeReactISurfGrid::grow_tally()
 bigint ComputeReactISurfGrid::memory_usage()
 {
   bigint bytes = 0;
-  bytes += ntotal*maxgrid * sizeof(double);     // array_grid
-  bytes += ntotal*maxtally * sizeof(double);    // array_surf_tally
+  bytes += (bigint) ntotal*maxgrid * sizeof(double);     // array_grid
+  bytes += (bigint) ntotal*maxtally * sizeof(double);    // array_surf_tally
   bytes += maxtally * sizeof(surfint);          // tally2surf
   return bytes;
 }
